pyproject.toml
[tool.poetry]
name = "cis-benchmark-bash-to-python"
version = "0.1.0"
description = ""
authors = ["gatewaynode <44749714+gatewaynode@users.noreply.github.com>"]
readme = "README.md"
packages = [{ include = "cis_benchmark_bash_to_python" }]

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.8"
pyyaml = "^6.0.2"
fabric = "^3.2.2"

[tool.poetry.scripts]
cis = "cis_benchmark_bash_to_python.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
